Modify instructions to specify Kubernetes 1.15
Instructions no longer work on GCP for 1.14. These require alpha
features to be enabled, but there is a conflict with alpha features that
causes the API Server to generate an incomplete swagger.json.

Surfacing CRDs in the OpenAPI Spec is beta in 1.15 and automatically
activated, so using 1.15 works.

DOES NOT WORK IN GCP WITH ALPHA FEATURES ENABLED, FOR ANY VERSION.