Add Rename helper, along with ref visitors
Rename will rename / renamespace an object, but will also update all
references to it.

We introduce the idea of a Ref and functions to discover and visit
references.

These are currently backed by a hard-coded list of kinds and ref
fields; we should move this to something backed by OpenAPI in future.